+++ CHANGELOG.md
@@ -3,6 +3,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
 ## Unreleased
 - Add a monitor to watch for changes in census county values
+- Add a script and manage.py command to automate county data collection
 
 ## 0.9.92 - 2016-12-31
 - 2017 update for county-level mortgage-limit data